Adds support for caching
