+++ functions/serializer.py
@@ -17,7 +17,6 @@ import codecs
 from functions_pb2 import *
 from io import BytesIO
 import numpy as np
-from shared import *
 
 SER_FORMAT = 'raw_unicode_escape'
 
@@ -100,18 +99,4 @@ def get_serializer(kind):
         return default_ser
     else:
         return default_ser
-
-def serialize_val(val, valobj=None):
-    if not valobj:
-        valobj = Value()
-
-    if isinstance(val, FluentFuture):
-        valobj.body = default_ser.dump(FluentReference(val.obj_id, True))
-    elif isinstance(val, np.ndarray):
-        valobj.body = numpy_ser.dump(val)
-        valobj.type = NUMPY
-    else:
-        valobj.body = default_ser.dump(val)
-
-    return valobj
